+++ tests/framework/microvm.py
@@ -1095,6 +1095,7 @@ class Microvm:
         rename_interfaces: dict = None,
         vsock_override: str = None,
         clock_realtime: bool = False,
+        pmem_overrides: list = None,
         *,
         uffd_handler_name: str = None,
     ):
@@ -1159,6 +1160,9 @@ class Microvm:
 
         if clock_realtime:
             optional_kwargs["clock_realtime"] = clock_realtime
+
+        if pmem_overrides is not None:
+            optional_kwargs["pmem_overrides"] = pmem_overrides
 
         self.api.snapshot_load.put(
             mem_backend=mem_backend,